requirements.txt
pyrogram==2.0.106
py-tgcalls>=2.2.0
TgCrypto==1.2.5
yt-dlp>=2024.8.0
aiohttp>=3.9.0
aiofiles>=23.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
mutagen>=1.47.0
Pillow>=10.0.0
